Build the 'ankr' table defined in the input file
It uses the syntax that is an extension to the kerning input file, but
the specifically named anchors as used here, although convenient to
reference from the state tables, they're not without problems.

First, they do not translate to the table data as-is, a feature found
throughout the input files in OS X Font Tools. Second, because both the
anchor and kerning tables have to be included in one input file, both
will have to be rebuild each time one is changed. Third, there's no way
to reuse actions that would otherwise refer to the same exact anchor
indices regardless of the glyph classes.